reorganize and homogenize some of the source classes, move some to a boneyard
