+++ prospect/sources/__init__.py
@@ -1,9 +1,11 @@
 from .galaxy_basis import *
-from .star_basis import *
 from .ssp_basis import *
+from .star_basis import *
 from .dust_basis import *
+from .boneyard import StepSFHBasis
 
-__all__ = ["StellarPopBasis", "CSPBasis",
+__all__ = ["CSPSpecBasis", "CSPBasis", "to_cgs",
+           "SSPBasis", "FastSSPBasis", "FastStepBasis",
+           "StepSFHBasis",
            "StarBasis", "BigStarBasis",
-           "SSPBasis", "FastSSPBasis", "StepSFHBasis", "CompositeSFH",
-           "BlackBodyDustBasis", "to_cgs"]
+           "BlackBodyDustBasis"]
